requirements.txt
amqp==1.4.9
ansible==2.4.0.0
appdirs==1.4.3
asn1crypto==0.24.0
attrs==17.1.0
Babel==2.3.4
bcrypt==3.1.4
billiard==3.3.0.23
cache-requests==4.0.0
celery==3.1.24
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
configparser==3.5.0
coreapi==2.3.3
cryptography==2.2
django-bootstrap3==7.1.0
django-celery==3.1.17
django-nose==1.4.5
djangorestframework==3.7.7
docopt==0.6.2
flake8==3.3.0
flower==0.9.1
futures==3.0.3
gnureadline==6.3.8
grafanalib==0.1.2
gunicorn==19.6.0
idna==2.6
influxdb==4.1.0
ipython==3.2.1
jedi==0.11.1
Jinja2==2.9.6
jsonschema==2.5.1
kombu==3.0.37
MarkupSafe==1.0
mccabe==0.6.1
multi-key-dict==2.0.3
packaging==16.8
paramiko==2.3.1
parso==0.1.1
pbr==1.10.0
pew==1.1.2
pipenv==11.1.3
prompt-toolkit==1.0.15
psutil==5.2.2
ptpython==0.41
py==1.4.33
pyasn1==0.3.7
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.5.0
Pygments==2.2.0
pymongo==3.4.0
PyNaCl==1.1.2
pyparsing==2.1.10
pyreadline==2.1
pytest==2.7.2
pytest-disable==0.2
python-dateutil==2.6.0
pytz==2016.7
PyYAML==3.12
redis==2.10.5
redislite==3.2.311
requests==2.18.4
six==1.11.0
uiautomator==0.1.35
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.3.0
wcwidth==0.1.7
websocket-client==0.37.0
yolk==0.4.3
